# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sulocraft API
  description: Backend service for the Sulocraft direct-to-consumer crochet storefront.
  version: 0.2.0
paths:
...
  /api/v1/cart:
    get:
      tags:
      - cart
      summary: Get Cart
      description: Retrieve current shopping cart. Returns empty cart representation
        if no cart exists.
      operationId: get_cart_api_v1_cart_get
      parameters:
      - name: x-cart-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Cart-Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: guest_cart_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Cart Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - cart
      summary: Clear Cart
      description: Clear all items from the current cart.
      operationId: clear_cart_api_v1_cart_delete
      parameters:
      - name: x-cart-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Cart-Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: guest_cart_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Cart Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/addresses:
    get:
      tags:
      - orders
      summary: List Addresses
      description: Retrieve saved delivery addresses for the authenticated customer.
      operationId: list_addresses_api_v1_addresses_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressOut'
                title: Response List Addresses Api V1 Addresses Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - orders
      summary: Create Address
      description: Save a new delivery address for the authenticated customer.
      operationId: create_address_api_v1_addresses_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/orders:
    post:
      tags:
      - orders
      summary: Checkout
      description: 'Convert the current shopping cart into a placed order.


        Snapshots products, variant details, prices, and shipping address.

        Decrements inventory, clears active cart items, and logs initial status history.'
      operationId: checkout_api_v1_orders_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: x-cart-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Cart-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      - name: guest_cart_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Cart Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - orders
      summary: List Orders
      description: Retrieve order history for the authenticated customer.
      operationId: list_orders_api_v1_orders_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderOut'
                title: Response List Orders Api V1 Orders Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/orders:
    get:
      tags:
      - admin
      summary: List Admin Orders
      description: List all orders across customers with status filters and search.
      operationId: list_admin_orders_api_v1_admin_orders_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by OrderStatus
          title: Status
        description: Filter by OrderStatus
      - name: paymentStatus
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by PaymentStatus
          title: Paymentstatus
        description: Filter by PaymentStatus
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search by orderNumber, customer name, email, or phone
          title: Q
        description: Search by orderNumber, customer name, email, or phone
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Pagesize
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /static/images/{image_path}:
    head:
      tags:
      - system
      summary: Get Mock Static Image
      description: Serve local static assets or redirect to high-res photography during
        development.
      operationId: get_mock_static_image_static_images__image_path__head
      parameters:
      - name: image_path
        in: path
        required: true
        schema:
          type: string
          title: Image Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - system
      summary: Get Mock Static Image
      description: Serve local static assets or redirect to high-res photography during
        development.
      operationId: get_mock_static_image_static_images__image_path__head
      parameters:
      - name: image_path
        in: path
        required: true
        schema:
          type: string
          title: Image Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    AddressCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 2
          title: Name
          description: Recipient full name
        phone:
          type: string
          maxLength: 20
          minLength: 10
          title: Phone
          description: 10-digit contact mobile number
        line1:
          type: string
          maxLength: 255
          minLength: 3
          title: Line1
          description: House/Flat number, building name, street
        line2:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Line2
          description: Apartment, suite, unit, etc.
        landmark:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Landmark
          description: Nearby landmark
        city:
          type: string
          maxLength: 100
          minLength: 2
          title: City
          description: City/District
        state:
          type: string
          maxLength: 100
          minLength: 2
          title: State
          description: State/Province
        postalCode:
          type: string
          maxLength: 10
          minLength: 6
          title: Postalcode
          description: 6-digit Indian PIN code
        country:
          type: string
          title: Country
          description: ISO 3166-1 alpha-2 country code
          default: IN
        isDefault:
          type: boolean
          title: Isdefault
          description: Set as default delivery address
          default: false
      type: object
      required:
      - name
      - phone
      - line1
      - city
      - state
      - postalCode
      title: AddressCreate
      description: Schema to create a new address.
    AddressOut:
      properties:
        id:
          type: integer
          title: Id
        userId:
          type: integer
          title: Userid
        name:
          type: string
          title: Name
        phone:
          type: string
          title: Phone
        line1:
          type: string
          title: Line1
        line2:
          anyOf:
          - type: string
          - type: 'null'
          title: Line2
        landmark:
          anyOf:
          - type: string
          - type: 'null'
          title: Landmark
        city:
          type: string
          title: City
        state:
          type: string
          title: State
        postalCode:
          type: string
          title: Postalcode
        country:
          type: string
          title: Country
          default: IN
        isDefault:
          type: boolean
          title: Isdefault
          default: false
        createdAt:
          type: string
          title: Createdat
      type: object
      required:
      - id
      - userId
      - name
      - phone
      - line1
      - city
      - state
      - postalCode
      - createdAt
      title: AddressOut
      description: Address output representation.
...
    AdminOrderDetailOut:
      properties:
        id:
          type: integer
          title: Id
        orderNumber:
          type: string
          title: Ordernumber
        customerName:
          type: string
          title: Customername
        customerPhone:
          type: string
          title: Customerphone
        customerEmail:
          anyOf:
          - type: string
          - type: 'null'
          title: Customeremail
        status:
          type: string
          title: Status
        paymentStatus:
          type: string
          title: Paymentstatus
        paymentMethod:
          type: string
          title: Paymentmethod
        currency:
          type: string
          title: Currency
          default: INR
        subtotal:
          type: integer
          title: Subtotal
        subtotalPaise:
          type: integer
          title: Subtotalpaise
        shippingFee:
          type: integer
          title: Shippingfee
        shippingFeePaise:
          type: integer
          title: Shippingfeepaise
        discountAmount:
          type: integer
          title: Discountamount
        discountAmountPaise:
          type: integer
          title: Discountamountpaise
        taxAmount:
          type: integer
          title: Taxamount
          default: 0
        taxAmountPaise:
          type: integer
          title: Taxamountpaise
          default: 0
        totalAmount:
          type: integer
          title: Totalamount
        totalAmountPaise:
          type: integer
          title: Totalamountpaise
        shippingAddress:
          additionalProperties: true
          type: object
          title: Shippingaddress
        billingAddress:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Billingaddress
        trackingNumber:
          anyOf:
          - type: string
          - type: 'null'
          title: Trackingnumber
        courierName:
          anyOf:
          - type: string
          - type: 'null'
          title: Couriername
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        items:
          items:
            $ref: '#/components/schemas/AdminOrderItemOut'
          type: array
          title: Items
        statusHistory:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Statushistory
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - orderNumber
      - customerName
      - customerPhone
      - status
      - paymentStatus
      - paymentMethod
      - subtotal
      - subtotalPaise
      - shippingFee
      - shippingFeePaise
      - discountAmount
      - discountAmountPaise
      - totalAmount
      - totalAmountPaise
      - shippingAddress
      - createdAt
      - updatedAt
      title: AdminOrderDetailOut
      description: Full order view for admin operations.
    AdminOrderItemOut:
      properties:
        id:
          type: integer
          title: Id
        productId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Productid
        variantId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Variantid
        productName:
          type: string
          title: Productname
        productSlug:
          anyOf:
          - type: string
          - type: 'null'
          title: Productslug
        sku:
          type: string
          title: Sku
        variantName:
          type: string
          title: Variantname
        productImage:
          anyOf:
          - type: string
          - type: 'null'
          title: Productimage
        unitPrice:
          type: integer
          title: Unitprice
        unitPricePaise:
          type: integer
          title: Unitpricepaise
        quantity:
          type: integer
          title: Quantity
        lineTotal:
          type: integer
          title: Linetotal
        lineTotalPaise:
          type: integer
          title: Linetotalpaise
        personalization:
          additionalProperties: true
          type: object
          title: Personalization
      type: object
      required:
      - id
      - productName
      - sku
      - variantName
      - unitPrice
      - unitPricePaise
      - quantity
      - lineTotal
      - lineTotalPaise
      title: AdminOrderItemOut
      description: Snapshotted order item for admin view.
    AdminOrderListOut:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AdminOrderDetailOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        pageSize:
          type: integer
          title: Pagesize
      type: object
      required:
      - items
      - total
      - page
      - pageSize
      title: AdminOrderListOut
      description: Paginated list of orders for administration.
...
    CartItemOut:
      properties:
        id:
          type: integer
          title: Id
        productId:
          type: integer
          title: Productid
        productName:
          type: string
          title: Productname
        productSlug:
          type: string
          title: Productslug
        productImage:
          type: string
          title: Productimage
        variantId:
          type: integer
          title: Variantid
        variantSku:
          type: string
          title: Variantsku
        variantName:
          type: string
          title: Variantname
        unitPrice:
          type: integer
          title: Unitprice
        unitPricePaise:
          anyOf:
          - type: integer
          - type: 'null'
          title: Unitpricepaise
        currency:
          type: string
          title: Currency
          default: INR
        compareAtPrice:
          anyOf:
          - type: integer
          - type: 'null'
          title: Compareatprice
        quantity:
          type: integer
          title: Quantity
        lineTotal:
          type: integer
          title: Linetotal
        lineTotalPaise:
          anyOf:
          - type: integer
          - type: 'null'
          title: Linetotalpaise
        personalization:
          additionalProperties: true
          type: object
          title: Personalization
        stockAvailable:
          type: integer
          title: Stockavailable
      type: object
      required:
      - id
      - productId
      - productName
      - productSlug
      - productImage
      - variantId
      - variantSku
      - variantName
      - unitPrice
      - quantity
      - lineTotal
      - stockAvailable
      title: CartItemOut
      description: Representation of an individual cart line item.
...
    CartOut:
      properties:
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        guestToken:
          anyOf:
          - type: string
          - type: 'null'
          title: Guesttoken
        items:
          items:
            $ref: '#/components/schemas/CartItemOut'
          type: array
          title: Items
          default: []
        itemCount:
          type: integer
          title: Itemcount
          default: 0
        subtotal:
          type: integer
          title: Subtotal
          default: 0
        subtotalPaise:
          anyOf:
          - type: integer
          - type: 'null'
          title: Subtotalpaise
        currency:
          type: string
          title: Currency
          default: INR
        status:
          type: string
          title: Status
          default: ACTIVE
      type: object
      title: CartOut
      description: Complete shopping cart representation.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrderCreate:
      properties:
        addressId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Addressid
          description: ID of saved address to use
        shippingAddress:
          anyOf:
          - $ref: '#/components/schemas/AddressCreate'
          - type: 'null'
          description: Inline address if not using saved address
        paymentMethod:
          type: string
          title: Paymentmethod
          description: 'Payment method: COD, UPI, CARD, NETBANKING'
          default: COD
        customerName:
          anyOf:
          - type: string
          - type: 'null'
          title: Customername
          description: Customer full name
        customerPhone:
          anyOf:
          - type: string
          - type: 'null'
          title: Customerphone
          description: Customer mobile number
        customerEmail:
          anyOf:
          - type: string
          - type: 'null'
          title: Customeremail
          description: Customer email address
        couponCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Couponcode
          description: Optional promo coupon code
        notes:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Notes
          description: Optional delivery or gift notes
      type: object
      title: OrderCreate
      description: Order checkout request.
    OrderItemOut:
      properties:
        id:
          type: integer
          title: Id
        orderId:
          type: integer
          title: Orderid
        productId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Productid
        productVariantId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Productvariantid
        productName:
          type: string
          title: Productname
        productSlug:
          anyOf:
          - type: string
          - type: 'null'
          title: Productslug
        sku:
          type: string
          title: Sku
        variantName:
          type: string
          title: Variantname
        productImage:
          anyOf:
          - type: string
          - type: 'null'
          title: Productimage
        unitPrice:
          type: integer
          title: Unitprice
        unitPricePaise:
          type: integer
          title: Unitpricepaise
        quantity:
          type: integer
          title: Quantity
        lineTotal:
          type: integer
          title: Linetotal
        lineTotalPaise:
          type: integer
          title: Linetotalpaise
        personalization:
          additionalProperties: true
          type: object
          title: Personalization
      type: object
      required:
      - id
      - orderId
      - productName
      - sku
      - variantName
      - unitPrice
      - unitPricePaise
      - quantity
      - lineTotal
      - lineTotalPaise
      title: OrderItemOut
      description: Snapshotted line item of an order.
    OrderOut:
      properties:
        id:
          type: integer
          title: Id
        orderNumber:
          type: string
          title: Ordernumber
        userId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Userid
        customerName:
          type: string
          title: Customername
        customerPhone:
          type: string
          title: Customerphone
        customerEmail:
          anyOf:
          - type: string
          - type: 'null'
          title: Customeremail
        shippingAddress:
          additionalProperties: true
          type: object
          title: Shippingaddress
        billingAddress:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Billingaddress
        status:
          type: string
          title: Status
        paymentStatus:
          type: string
          title: Paymentstatus
        paymentMethod:
          type: string
          title: Paymentmethod
        currency:
          type: string
          title: Currency
          default: INR
        subtotal:
          type: integer
          title: Subtotal
        subtotalPaise:
          type: integer
          title: Subtotalpaise
        shippingFee:
          type: integer
          title: Shippingfee
        shippingFeePaise:
          type: integer
          title: Shippingfeepaise
        discountAmount:
          type: integer
          title: Discountamount
          default: 0
        discountAmountPaise:
          type: integer
          title: Discountamountpaise
          default: 0
        taxAmount:
          type: integer
          title: Taxamount
          default: 0
        taxAmountPaise:
          type: integer
          title: Taxamountpaise
          default: 0
        totalAmount:
          type: integer
          title: Totalamount
        totalAmountPaise:
          type: integer
          title: Totalamountpaise
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        items:
          items:
            $ref: '#/components/schemas/OrderItemOut'
          type: array
          title: Items
          default: []
        statusHistory:
          items:
            $ref: '#/components/schemas/OrderStatusHistoryOut'
          type: array
          title: Statushistory
        trackingNumber:
          anyOf:
          - type: string
          - type: 'null'
          title: Trackingnumber
        courierName:
          anyOf:
          - type: string
          - type: 'null'
          title: Couriername
        estimatedDelivery:
          anyOf:
          - type: string
          - type: 'null'
          title: Estimateddelivery
        createdAt:
          type: string
          title: Createdat
        updatedAt:
          type: string
          title: Updatedat
      type: object
      required:
      - id
      - orderNumber
      - customerName
      - customerPhone
      - shippingAddress
      - status
      - paymentStatus
      - paymentMethod
      - subtotal
      - subtotalPaise
      - shippingFee
      - shippingFeePaise
      - totalAmount
      - totalAmountPaise
      - createdAt
      - updatedAt
      title: OrderOut
      description: Detailed order response representation.
    OrderStatusHistoryOut:
      properties:
        id:
          type: integer
          title: Id
        status:
          type: string
          title: Status
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        timestamp:
          type: string
          title: Timestamp
      type: object
      required:
      - id
      - status
      - timestamp
      title: OrderStatusHistoryOut
      description: Audit log entry for order status transition.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...